# fluid-remote/docs/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: API for managing virtual machine sandboxes using libvirt
  title: fluid-remote API
  version: 0.0.1-beta
servers:
  - url: /
...
paths:
...
  /v1/ansible/playbooks:
    get:
      description: Lists all Ansible playbooks
      operationId: listPlaybooks
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_ansible.listPlaybooksResponse"
          description: OK
      summary: List playbooks
      tags:
        - Ansible Playbooks
    post:
      description: Creates a new Ansible playbook
      operationId: createPlaybook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_ansible.createPlaybookRequest"
        description: Playbook creation parameters
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_ansible.createPlaybookResponse"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Conflict
      summary: Create playbook
      tags:
        - Ansible Playbooks
      x-codegen-request-body-name: request
  /v1/ansible/playbooks/{playbook_name}:
    delete:
      description: Deletes a playbook and all its tasks
      operationId: deletePlaybook
      parameters:
        - description: Playbook name
          in: path
          name: playbook_name
          required: true
          schema:
            type: string
      responses:
        "204":
          content: {}
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Not Found
      summary: Delete playbook
      tags:
        - Ansible Playbooks
    get:
      description: Gets a playbook and its tasks by name
      operationId: getPlaybook
      parameters:
        - description: Playbook name
          in: path
          name: playbook_name
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_ansible.getPlaybookResponse"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Not Found
      summary: Get playbook
      tags:
        - Ansible Playbooks
  /v1/ansible/playbooks/{playbook_name}/export:
    get:
      description: Exports a playbook as raw YAML
      operationId: exportPlaybook
      parameters:
        - description: Playbook name
          in: path
          name: playbook_name
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_ansible.exportPlaybookResponse"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Not Found
      summary: Export playbook
      tags:
        - Ansible Playbooks
  /v1/ansible/playbooks/{playbook_name}/tasks:
    post:
      description: Adds a new task to an existing playbook
      operationId: addPlaybookTask
      parameters:
        - description: Playbook name
          in: path
          name: playbook_name
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_ansible.addTaskRequest"
        description: Task parameters
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_ansible.addTaskResponse"
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Not Found
      summary: Add task to playbook
      tags:
        - Ansible Playbooks
      x-codegen-request-body-name: request
  /v1/ansible/playbooks/{playbook_name}/tasks/{task_id}:
    delete:
      description: Removes a task from a playbook
      operationId: deletePlaybookTask
      parameters:
        - description: Playbook name
          in: path
          name: playbook_name
          required: true
          schema:
            type: string
        - description: Task ID
          in: path
          name: task_id
          required: true
          schema:
            type: string
      responses:
        "204":
          content: {}
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Not Found
      summary: Delete task
      tags:
        - Ansible Playbooks
    put:
      description: Updates an existing task in a playbook
      operationId: updatePlaybookTask
      parameters:
        - description: Playbook name
          in: path
          name: playbook_name
          required: true
          schema:
            type: string
        - description: Task ID
          in: path
          name: task_id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_ansible.updateTaskRequest"
        description: Task update parameters
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_ansible.updateTaskResponse"
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Not Found
      summary: Update task
      tags:
        - Ansible Playbooks
      x-codegen-request-body-name: request
  /v1/ansible/playbooks/{playbook_name}/tasks/reorder:
    patch:
      description: Reorders tasks in a playbook
      operationId: reorderPlaybookTasks
      parameters:
        - description: Playbook name
          in: path
          name: playbook_name
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_ansible.reorderTasksRequest"
        description: New task order
        required: true
      responses:
        "204":
          content: {}
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/fluid-remote_internal_error.ErrorResponse"
          description: Not Found
      summary: Reorder tasks
      tags:
        - Ansible Playbooks
      x-codegen-request-body-name: request
...
components:
  schemas:
...
    internal_ansible.addTaskRequest:
      properties:
        module:
          type: string
        name:
          type: string
        params:
          properties: {}
          type: object
      type: object
    internal_ansible.addTaskResponse:
      example:
        task:
          module: module
          name: name
          created_at: created_at
          id: id
          position: 0
          params: "{}"
          playbook_id: playbook_id
      properties:
        task:
          $ref: "#/components/schemas/fluid-remote_internal_store.PlaybookTask"
      type: object
    internal_ansible.createPlaybookRequest:
      properties:
        become:
          type: boolean
        hosts:
          type: string
        name:
          type: string
      type: object
    internal_ansible.createPlaybookResponse:
      example:
        playbook:
          file_path: file_path
          updated_at: updated_at
          hosts: hosts
          name: name
          created_at: created_at
          id: id
          become: true
      properties:
        playbook:
          $ref: "#/components/schemas/fluid-remote_internal_store.Playbook"
      type: object
    internal_ansible.exportPlaybookResponse:
      example:
        yaml: yaml
      properties:
        yaml:
          type: string
      type: object
    internal_ansible.getPlaybookResponse:
      example:
        tasks:
          - module: module
            name: name
            created_at: created_at
            id: id
            position: 0
            params: "{}"
            playbook_id: playbook_id
          - module: module
            name: name
            created_at: created_at
            id: id
            position: 0
            params: "{}"
            playbook_id: playbook_id
        playbook:
          file_path: file_path
          updated_at: updated_at
          hosts: hosts
          name: name
          created_at: created_at
          id: id
          become: true
      properties:
        playbook:
          $ref: "#/components/schemas/fluid-remote_internal_store.Playbook"
        tasks:
          items:
            $ref: "#/components/schemas/fluid-remote_internal_store.PlaybookTask"
          type: array
      type: object
    internal_ansible.listPlaybooksResponse:
      example:
        total: 0
        playbooks:
          - file_path: file_path
            updated_at: updated_at
            hosts: hosts
            name: name
            created_at: created_at
            id: id
            become: true
          - file_path: file_path
            updated_at: updated_at
            hosts: hosts
            name: name
            created_at: created_at
            id: id
            become: true
      properties:
        playbooks:
          items:
            $ref: "#/components/schemas/fluid-remote_internal_store.Playbook"
          type: array
        total:
          type: integer
      type: object
    internal_ansible.reorderTasksRequest:
      properties:
        task_ids:
          items:
            type: string
          type: array
      type: object
    internal_ansible.updateTaskRequest:
      properties:
        module:
          type: string
        name:
          type: string
        params:
          properties: {}
          type: object
      type: object
    internal_ansible.updateTaskResponse:
      example:
        task:
          module: module
          name: name
          created_at: created_at
          id: id
          position: 0
          params: "{}"
          playbook_id: playbook_id
      properties:
        task:
          $ref: "#/components/schemas/fluid-remote_internal_store.PlaybookTask"
      type: object
...
    fluid-remote_internal_error.ErrorResponse:
      example:
        code: 0
        details: details
        error: error
      properties:
        code:
          type: integer
        details:
          type: string
        error:
          type: string
      type: object
...
    fluid-remote_internal_store.Playbook:
      example:
        file_path: file_path
        updated_at: updated_at
        hosts: hosts
        name: name
        created_at: created_at
        id: id
        become: true
      properties:
        become:
          description: whether to use privilege escalation
          type: boolean
        created_at:
          type: string
        file_path:
          description: rendered YAML file path
          type: string
        hosts:
          description: 'target hosts pattern (e.g., "all", "webservers")'
          type: string
        id:
          type: string
        name:
          description: unique playbook name
          type: string
        updated_at:
          type: string
      type: object
    fluid-remote_internal_store.PlaybookTask:
      example:
        module: module
        name: name
        created_at: created_at
        id: id
        position: 0
        params: "{}"
        playbook_id: playbook_id
      properties:
        created_at:
          type: string
        id:
          type: string
        module:
          description: "ansible module (apt, shell, copy, etc.)"
          type: string
        name:
          description: task name/description
          type: string
        params:
          description: module-specific parameters
          properties: {}
          type: object
        playbook_id:
          type: string
        position:
          description: ordering within playbook
          type: integer
      type: object
...